{3a7ef96c-cfea-4535-epoc-165289847649}

+++ widgets/batch/drive.bat
@@ -21,15 +21,15 @@ SET thisSelected=
 SET thisAction=
 SET thisView=
 IF NOT [%2] == [/v] (
-		SET thisSelected=%2
-	) ELSE (
-		SET thisView=true
-		IF NOT [%3] == [] (SET thisSelected=%3)
-	)
+        SET thisSelected=%2
+    ) ELSE (
+        SET thisView=true
+        IF NOT [%3] == [] (SET thisSelected=%3)
+    )
 IF NOT [%1] == [] (
-		SET thisAction=%1
-		CALL :GO
-	)
+        SET thisAction=%1
+        CALL :GO
+    )
 GOTO:EOF
 :DRIVESCAN
 IF EXIST "D:\" SET thisDrive=B & CALL :SETDRIVE
@@ -63,10 +63,10 @@ GOTO:EOF
 :SETDRIVE
 SET driveIDPath=%thisDrive:~0,1%:\drive.id.sys
 IF EXIST "%driveIDPath%" (
-		SET /p driveID=<"%driveIDPath%"
-	) else (
-		CALL :driveIDPathNOTEXIST
-	)
+        SET /p driveID=<"%driveIDPath%"
+    ) else (
+        CALL :driveIDPathNOTEXIST
+    )
 CALL timestamp default noEcho
 SET thisDMR=%manageDrivesRoot%\%driveID%
 SET thisDMRLabel=%thisDMR%\label.txt
@@ -75,45 +75,45 @@ SET thisDMRLetter=%thisDMR%\letter.txt
 SET thisDMRIndex=%thisDMR%\indexSettings.txt
 SET thisDMRPriority=%thisDMR%\prioritySettings.txt
 IF EXIST "%thisDMR%" (
-		SET /p driveLabel=<"%thisDMRLabel%"
-		echo %now%,%thisDrive:~0,1%,Registered>> %thisDMRLog%
-		echo %thisDrive:~0,1%>"%thisDMRLetter%"
-		IF NOT EXIST "%thisDMRIndex%" (echo F>"%thisDMRIndex%")
-		IF NOT EXIST "%thisDMRPriority%" (echo 20>"%thisDMRPriority%")
-		ping 127.0.0.1 -n 1 > nul
-		SET /p thisIndexSetting=<"%thisDMRIndex%"
-		SET /p thisPrioritySetting=<"%thisDMRPriority%"
-	) ELSE (
-		mkdir %thisDMR%
-		ping 127.0.0.1 -n 1 > nul
-		IF EXIST "%thisDMR%" (
-				echo %now%,Created>> %thisDMRLog%
-				echo %now%,%thisDrive:~0,1%,Registered>> %thisDMRLog%
-				echo N>"%thisDMRIndex%"
-				echo 20>"%thisDMRPriority%"
-				echo %thisDrive:~0,1%>"%thisDMRLetter%"
-				echo %thisDrive:~0,1%:\>"%thisDMRLabel%"
-				notepad "%thisDMRLabel%"
-				ping 127.0.0.1 -n 1 > nul
-				SET /p driveLabel=<"%thisDMRLabel%"
-				SET /p thisIndexSetting=<"%thisDMRIndex%"
-				SET /p thisPrioritySetting=<"%thisDMRPriority%"
-			)
-	)
+        SET /p driveLabel=<"%thisDMRLabel%"
+        echo %now%,%thisDrive:~0,1%,Registered>> %thisDMRLog%
+        echo %thisDrive:~0,1%>"%thisDMRLetter%"
+        IF NOT EXIST "%thisDMRIndex%" (echo F>"%thisDMRIndex%")
+        IF NOT EXIST "%thisDMRPriority%" (echo 20>"%thisDMRPriority%")
+        ping 127.0.0.1 -n 1 > nul
+        SET /p thisIndexSetting=<"%thisDMRIndex%"
+        SET /p thisPrioritySetting=<"%thisDMRPriority%"
+    ) ELSE (
+        mkdir %thisDMR%
+        ping 127.0.0.1 -n 1 > nul
+        IF EXIST "%thisDMR%" (
+                echo %now%,Created>> %thisDMRLog%
+                echo %now%,%thisDrive:~0,1%,Registered>> %thisDMRLog%
+                echo N>"%thisDMRIndex%"
+                echo 20>"%thisDMRPriority%"
+                echo %thisDrive:~0,1%>"%thisDMRLetter%"
+                echo %thisDrive:~0,1%:\>"%thisDMRLabel%"
+                notepad "%thisDMRLabel%"
+                ping 127.0.0.1 -n 1 > nul
+                SET /p driveLabel=<"%thisDMRLabel%"
+                SET /p thisIndexSetting=<"%thisDMRIndex%"
+                SET /p thisPrioritySetting=<"%thisDMRPriority%"
+            )
+    )
 
-echo          %thisDrive:~0,1%:\ %driveLabel%	%thisIndexSetting%	%thisPrioritySetting%	%driveID%
+echo          %thisDrive:~0,1%:\ %driveLabel%    %thisIndexSetting%    %thisPrioritySetting%    %driveID%
 GOTO:EOF
 
 :driveIDPathNOTEXIST
-		SET driveIDtmpFile=tmp001X.txt
-		echo '%driveIDtmpFile%'
-		CALL %scriptroot%\genguid.bat > %driveIDtmpFile%
-		ping 127.0.0.1 -n 1 > nul
-		SET /p driveID=<%driveIDtmpFile%
-		del /q %driveIDtmpFile%
-		echo %driveID%>"%driveIDPath%"
-		ping 127.0.0.1 -n 1 > nul
-		IF EXIST "%driveIDPath%" (attrib +h "%driveIDPath%")
+        SET driveIDtmpFile=tmp001X.txt
+        echo '%driveIDtmpFile%'
+        CALL %scriptroot%\genguid.bat > %driveIDtmpFile%
+        ping 127.0.0.1 -n 1 > nul
+        SET /p driveID=<%driveIDtmpFile%
+        del /q %driveIDtmpFile%
+        echo %driveID%>"%driveIDPath%"
+        ping 127.0.0.1 -n 1 > nul
+        IF EXIST "%driveIDPath%" (attrib +h "%driveIDPath%")
 GOTO:EOF
 
 :VAR
@@ -134,36 +134,36 @@ echo %thisSelected%
 
 SET settingIndex=
 IF [%thisView%] == [] (
-		IF [%thisAction%] == [/i] CALL :indexSettingsChange
-		IF [%thisAction%] == [/index] CALL :indexSettingsChange
+        IF [%thisAction%] == [/i] CALL :indexSettingsChange
+        IF [%thisAction%] == [/index] CALL :indexSettingsChange
 
-		IF [%thisAction%] == [/p] CALL :prioritySettingsChange
-		IF [%thisAction%] == [/priority] CALL :prioritySettingsChange
-	)
+        IF [%thisAction%] == [/p] CALL :prioritySettingsChange
+        IF [%thisAction%] == [/priority] CALL :prioritySettingsChange
+    )
 CALL b back
 GOTO:EOF
 
 :indexSettingsChange
-	type indexSettings.txt
-	set /p settingIndex=Change Index Setting to: 
+    type indexSettings.txt
+    set /p settingIndex=Change Index Setting to: 
 
-	IF NOT [%settingIndex%] == [] (echo %settingIndex:~0,1%>indexSettings.txt) ELSE (echo ERROR)
-	type indexSettings.txt
+    IF NOT [%settingIndex%] == [] (echo %settingIndex:~0,1%>indexSettings.txt) ELSE (echo ERROR)
+    type indexSettings.txt
 GOTO:EOF
 
 :prioritySettingsChange
-	type prioritySettings.txt
-	set /p settingPriority=Change Priority Setting to: 
-	IF NOT [%settingPriority%] == [] (
-			CALL :prioritySettingsChangeSelected
-		) ELSE (
-			echo ERROR
-		)
-	type prioritySettings.txt
+    type prioritySettings.txt
+    set /p settingPriority=Change Priority Setting to: 
+    IF NOT [%settingPriority%] == [] (
+            CALL :prioritySettingsChangeSelected
+        ) ELSE (
+            echo ERROR
+        )
+    type prioritySettings.txt
 GOTO:EOF
 
 :prioritySettingsChangeSelected
-	echo %settingPriority%>prioritySettings.txt
-	IF EXIST "%mdrIndex%\*%thisSelected%*" (rename "%mdrIndex%\*%thisSelected%*" "%settingPriority:~0,2%%thisSelected%.txt")
+    echo %settingPriority%>prioritySettings.txt
+    IF EXIST "%mdrIndex%\*%thisSelected%*" (rename "%mdrIndex%\*%thisSelected%*" "%settingPriority:~0,2%%thisSelected%.txt")
 GOTO:EOF
